+++ src/app/api/auth/signin/page.tsx
@@ -1,5 +1,5 @@
 import { getServerSession } from "next-auth"
-import { authOptions } from "../[...nextauth]/route"
+import { authOptions } from "@/lib/auth"
 import { redirect } from "next/navigation"
 import SDApplicationBar from "@/components/SDApplicationBar"
 import Link from "next/link"
